Rename: customisable_nav_bar -> navbar
