--- front/app/modules/commercial/customisable_nav_bar/admin/CustomisableNavigationSettings.tsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default () => (
-  <div>Commercial version: navigation tab</div>
-);
