+++ front/app/modules/commercial/navbar/admin/CustomisableNavigationSettings.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export default () => <div>Commercial version: navigation tab</div>;
